Only apply SSE4 compiler flags to the code that needs it
This prevents the optimizer from making use of SSE4-specific optimizations
to code which will run before the availability of SSE4 has been verified.

Fixes #41.  Thanks to http://stackoverflow.com/a/32391337/1468366 by pmed
for the suggestion of how to formulate this split for gyp.
